Split TGA bits from u8x8_d_bitmap.c
This moves the TGA export code into a new file called
u8x8_d_bitmap_tga.c, which removes the stdio.h dependency from
u8x8_d_bitmap.c (making that file compilable on microcontroller targets
as well).

This slightly reworks the TGA export code so it uses the external bitmap
device API instead of reading directly from the global u8x8_bitmap
struct.

This also moves u8x8_d_bitmap.c into csrc, so it is included in all
builds (including the Arduino build).
