+++ csrc/u8x8_d_bitmap.c
@@ -9,7 +9,6 @@
 #include "stdlib.h"	/* malloc */
 #include "stdint.h"	/* uint16_t */
 #include "string.h"	/* memcpy */
-#include "stdio.h"	/* FILE */
 #include "u8g2.h"		/* because of u8g2_Setup... */
 
 /*========================================================*/
@@ -79,65 +78,6 @@ uint8_t u8x8_bitmap_GetPixel(u8x8_bitmap_t *b, uint16_t x, uint16_t y)
   return 1;
 }
 
-static void tga_write_byte(FILE *fp, uint8_t byte)
-{
-  fputc(byte, fp);
-}
-
-static void tga_write_word(FILE *fp, uint16_t word)
-{
-  tga_write_byte(fp, word&255);
-  tga_write_byte(fp, word>>8);
-}
-
-void u8x8_bitmap_SaveTGA(u8x8_bitmap_t *b, const char *name)
-{
-  FILE *fp;
-  uint16_t x, y;
-  
-  fp = fopen(name, "wb");
-  if ( fp != NULL )
-  {
-    tga_write_byte(fp, 0);		/* no ID */
-    tga_write_byte(fp, 0);		/* no color map */
-    tga_write_byte(fp, 2);		/* uncompressed true color */
-    tga_write_word(fp, 0);		
-    tga_write_word(fp, 0);		
-    tga_write_byte(fp, 0);		
-    tga_write_word(fp, 0);		/* x origin */
-    tga_write_word(fp, 0);		/* y origin */
-    tga_write_word(fp, b->pixel_width);		/* width */
-    tga_write_word(fp, b->pixel_height);		/* height */
-    tga_write_byte(fp, 24);		/* color depth */
-    tga_write_byte(fp, 0);	
-    for( y = 0; y < b->pixel_height; y++ )
-    {
-      for( x = 0; x < b->pixel_width; x++ )
-      {
-	if ( u8x8_bitmap_GetPixel(b, x, b->pixel_height-y-1) == 0 )
-	{
-	  tga_write_byte(fp, 255);		/* R */
-	  tga_write_byte(fp, 255);		/* G */
-	  tga_write_byte(fp, 255);		/* B */
-	}
-	else
-	{
-	  tga_write_byte(fp, 0);		/* R */
-	  tga_write_byte(fp, 0);		/* G */
-	  tga_write_byte(fp, 0);		/* B */
-	}
-      }
-    }
-    tga_write_word(fp, 0);
-    tga_write_word(fp, 0);
-    tga_write_word(fp, 0);
-    tga_write_word(fp, 0);
-    fwrite("TRUEVISION-XFILE.", 18, 1, fp);
-    fclose(fp);
-  }
-}
-
-
 /*========================================================*/
 /* global objects for the bitmap */
 
@@ -197,11 +137,6 @@ static void u8x8_DrawBitmapTiles(U8X8_UNUSED u8x8_t *u8x8, uint16_t tx, uint16_t
 uint8_t u8x8_GetBitmapPixel(U8X8_UNUSED u8x8_t *u8x8, uint16_t x, uint16_t y)
 {
   return u8x8_bitmap_GetPixel(&u8x8_bitmap, x, y);
-}
-
-void u8x8_SaveBitmapTGA(U8X8_UNUSED u8x8_t *u8x8, const char *filename)
-{
-  u8x8_bitmap_SaveTGA(&u8x8_bitmap, filename);
 }
 
 /*========================================================*/
